pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "optimized-llm-planning-memory"
version = "0.1.0"
description = "Context-compression training for LLM travel-planning agents via PPO"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Columbia MSDS STATGR5293 Team" }]

dependencies = [
    # Travel simulator (local sibling package — resolved via [tool.uv.sources])
    "travel-world",

    # Core data modelling
    "pydantic>=2.7",
    "pydantic-settings>=2.3",

    # LLM orchestration + provider portability
    "pydantic-ai>=0.0.15",
    "litellm>=1.40",
    "instructor>=1.3",

    # Reinforcement learning
    "gymnasium>=0.29",
    "stable-baselines3>=2.3",

    # ML / compressor model
    # torch dropped Intel Mac (x86_64) wheels starting from 2.3; 2.2.x is the last compatible release
    "torch>=2.3; sys_platform != 'darwin' or platform_machine != 'x86_64'",
    "torch>=2.2,<2.3; sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "transformers>=4.44",
    "peft>=0.11",        # LoRA adapters via HuggingFace PEFT
    "accelerate>=0.33",

    # HTTP (optional simulator transport, utilities)
    "httpx>=0.27",

    # Configuration composition
    "hydra-core>=1.3",
    "omegaconf>=2.3",

    # Logging + observability
    "structlog>=24.1",
    "tensorboard>=2.17",
    "rich>=13.7",        # pretty terminal output / visualization
]

[project.optional-dependencies]
wandb = ["wandb>=0.17"]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "respx>=0.21",       # mock httpx calls in tests
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "pytest-json-report>=1.5",  # structured JSON output for test runner
]
test-ui = [
    "streamlit>=1.37",
    "streamlit-autorefresh>=0.0.1",
    "pandas>=2.2",
]
notebook = [
    # Jupyter environment
    "jupyterlab>=4.2",
    "ipykernel>=6.29",
    # Notebook-only data / viz dependencies (not required by core package)
    "pandas>=2.2",
    "matplotlib>=3.9",
    "tqdm>=4.66",
]

[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
]

[tool.uv.sources]
# travel-world lives in a sibling directory; uv resolves it as an editable
# local install so changes to the simulator are immediately visible.
# Both repos must be cloned into the same parent directory (see README).
travel-world = { path = "../my-travel-world", editable = true }

[project.scripts]
run-episode    = "optimized_llm_planning_memory.scripts.run_episode:main"
run-training   = "optimized_llm_planning_memory.scripts.run_training:main"
run-evaluation = "optimized_llm_planning_memory.scripts.run_evaluation:main"

[tool.hatch.build.targets.wheel]
packages = ["src/optimized_llm_planning_memory"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"'); downloads HuggingFace weights",
    "integration: marks tests that require a live API key",
    "unit: fast isolated unit tests with all external deps mocked",
    "module_test: multi-class workflow tests within a single component",
    "system_test: cross-component end-to-end tests (LLM mocked, real logic)",
]
